+++ mapie/regression/regression.py
@@ -2,6 +2,8 @@ from __future__ import annotations
 
 from typing import Iterable, Optional, Tuple, Union, cast
 
+import warnings
+
 import numpy as np
 from sklearn.base import BaseEstimator, RegressorMixin
 from sklearn.linear_model import LinearRegression
@@ -532,6 +534,7 @@ class MapieRegressor(BaseEstimator, RegressorMixin):
         ensemble: bool = False,
         alpha: Optional[Union[float, Iterable[float]]] = None,
         optimize_beta: bool = False,
+        allow_infinite_bounds: bool = False,
     ) -> Union[NDArray, Tuple[NDArray, NDArray]]:
         """
         Predict target on new samples with confidence intervals.
@@ -576,6 +579,9 @@ class MapieRegressor(BaseEstimator, RegressorMixin):
 
             By default ``False``.
 
+        allow_infinite_bounds: bool
+            Allow infinite prediction intervals to be produced.
+
         Returns
         -------
         Union[NDArray, Tuple[NDArray, NDArray]]
@@ -598,14 +604,16 @@ class MapieRegressor(BaseEstimator, RegressorMixin):
 
         else:
             if optimize_beta and self.method != 'enbpi':
-                raise UserWarning(
-                    "Beta optimisation should only be used for "
-                    "method='enbpi'."
+                warnings.warn(
+                    "WARNING: Beta optimisation should only be used for "
+                    "method='enbpi'.",
+                    UserWarning
                 )
 
-            n = len(self.conformity_scores_)
             alpha_np = cast(NDArray, alpha)
-            check_alpha_and_n_samples(alpha_np, n)
+            if not allow_infinite_bounds:
+                n = len(self.conformity_scores_)
+                check_alpha_and_n_samples(alpha_np, n)
 
             y_pred, y_pred_low, y_pred_up = \
                 self.conformity_score_function_.get_bounds(